Add features 1. Protect multiple execution of daemon 2. Protect a error of client caused by terminating server 3. \'stop\' command

+++ taskmasterd.py
@@ -6,7 +6,7 @@
 #    By: ihwang <ihwang@student.hive.fi>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2020/08/25 03:24:46 by tango             #+#    #+#              #
-#    Updated: 2020/08/26 02:31:50 by ihwang           ###   ########.fr        #
+#    Updated: 2020/08/26 21:58:56 by ihwang           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -40,27 +40,43 @@ def stopsignal_handler(signum, frame):
                 p._pid.terminate()
 
 def listen_sighup(signum, frame):
+    global g_programs
+    global server
+    log.info("server: Detected %(sig)d signal has been sent", {"sig": signum})
     g_programs = create_configured_programs(server._config)
     execute.load_programs(g_programs)
+    execute.auto_start(server)
+
+def listen_sigterm(signum, frame):
+    log.info("server: Detected %(sig)d signal has been sent", {"sig": signum})
+    global g_programs
+    global server
+    for p in g_programs:
+        if p._stopsignal == signum and p._start_status == "Running":
+            log.info("server: %(n)s(%(p)s) will be terminated by daemon", {"n": p._name, "p": str(p._pid.pid)})
+            p._pid.terminate()
+            log.info("server: %(n)s(%(p)s) was killed", {"n": p._name, "p": str(p._pid.pid)})
+    log.info("server: cut the connection with client")
+    send_one_message(server._conn, "disconnect")
+    log.info("server: taskmaster daemon will be terminated")
+    exit(0)
 
 def listen_other_signals(signum, frame):
     log.info("server: Detected %(sig)d signal has been sent", {"sig": signum})
     global g_programs
     for p in g_programs:
-        if p._stopsignal == signum:
+        if p._stopsignal == signum and p._start_status == "Running":
             log.info("server: %(pid)d will be stoped after %(sec)d sec", {"pid": p._pid.pid, "sec": p._stoptime})
             try:
                 p._pid.wait(p._stoptime)
             except subprocess.TimeoutExpired:
                 p._pid.terminate()
-                log.info("server: %(pid)d \'%(n)s\' was killed", {"pid": p._pid.pid, "n": p._name})
-    if signum == signal.SIGTERM:
-        exit(0)
+                log.info("server: %(n)s(%(p)s) was killed", {"n": p._name, "p": str(p._pid.pid)})
 
 def listen_signals():
     signal.signal(signal.SIGHUP, listen_sighup)
+    signal.signal(signal.SIGTERM, listen_sigterm)
     signal.signal(signal.SIGINT, listen_other_signals)
-    signal.signal(signal.SIGTERM, listen_other_signals)
     signal.signal(signal.SIGQUIT, listen_other_signals)
 
 def main():
@@ -70,7 +86,7 @@ def main():
 
     go_daemon()
     server = Server()
-    server.start()
+    server.boot_server()
     g_programs = create_configured_programs(server._config)
     log.info("server: A config file has been loaded")
     execute = Execute(g_programs)
@@ -81,7 +97,7 @@ def main():
         execute.reload_status = False
         job = recv_one_message(server._conn)
         if job == "disconnected":
-            log.info("The client %(addr)s has been disconnected", {"addr": server._client_addr})
+            log.info("server: The client %(addr)s has been disconnected", {"addr": server._client_addr})
             server.wait_for_connect()
             g_programs = create_configured_programs(server._config)
             execute.load_programs(g_programs)
